requirements.txt
Pillow
langdetect
matplotlib
numpy
requests
torch
transliterate
unidecode
watchdog
opencv-python